==> Scripts/6-stconnect-library-sensitivity.R <==
## a311
## Carina Rauen Firkowski 
## February 28, 2024
##
## This script generates a stconnect library for sensitivity analyses on mean
## effective resistance distance for two parameters: 
##    - resistance map resolution
##    - number of sentinel nodes
## It is based on a single iteration and timestep, one protected area scenario, 
## and three species representative of diverse habitat needs and movement 
## abilities (American Marten, Red Backed Salamander, Red-breasted Nuthatch).
##
## The library is available to collaborators from SyncroSim Cloud at 
## 
## 
## NOTE: In order to run this script, please review line 53 to properly setup
##       the path to the Julia software on your machine.
##       Additionally, the paths to input data used in this script assume the 
##       working directory is set to "C:/gitprojects/A311". If setting a  
##       different working directory, make sure to update the following files:
##          - Initial Conditions Spatial Files.csv
##          - Transition Spatial Multipliers - Agnostic.csv
##          - Transition Spatial Multipliers - Zonation.csv
##          - Transition Spatial Multipliers - Prioritizr.csv



# Load constants
source("Scripts/0-constants.R")



# stconnect Library ------------------------------------------------------------

# Open a new SyncroSim session
stconnectSession <- session()

# Create a new stconnect Library
stconnectLibrary <- ssimLibrary(
  name = file.path(libraryDir, 
                   "a311-stconnect-sensitivity-resolution.ssim"),
  session = stconnectSession, overwrite = FALSE,
  package = "stconnect")

# Open default Project
stconnectProject <- rsyncrosim::project(ssimObject = stconnectLibrary, 
                                        project = "Definitions")


# Julia Configuration datasheet ------------------

# Set path to Julia executable
juliaConfiguration <- data.frame(Filename = "C:\\Users\\Administrator\\AppData\\Local\\Programs\\Julia-1.10.0\\bin\\julia.exe")

# Save datasheet
saveDatasheet(ssimObject = stconnectLibrary, 
              data = juliaConfiguration, 
              name = "stconnect_CCJuliaConfig")



# stconnect Project ------------------------------------------------------------

# Function to create or update sub-scenarios
edit_project <- function(projectInput, projectDatasheet){
  
  # Load input file
  inputDS <- read.csv(file.path(libraryInputsDir, projectInput))
  
  # Save datasheet
  saveDatasheet(ssimObject = stconnectProject, 
                data = inputDS, 
                name = projectDatasheet)
  
}

# Datasheet: Species
edit_project(projectInput = "Species.csv",
             projectDatasheet = "stconnect_Species")

# Datasheet: Terminology
edit_project(projectInput = "Terminology.csv",
             projectDatasheet = "stsim_Terminology")

# Datasheet: Ecoregion
edit_project(projectInput = "Ecoregion.csv",
             projectDatasheet = "stsim_Stratum")

# Datasheet: MRC
edit_project(projectInput = "MRC.csv",
             projectDatasheet = "stsim_SecondaryStratum")

# Datasheet: Terrain type
edit_project(projectInput = "TerrainType.csv",
             projectDatasheet = "stsim_TertiaryStratum")

# Datasheet: Class
edit_project(projectInput = "Class.csv",
             projectDatasheet = "stsim_StateLabelX")

# Datasheet: Subclass
edit_project(projectInput = "Subclass.csv",
             projectDatasheet = "stsim_StateLabelY")

# Datasheet: State class
edit_project(projectInput = "State Class.csv",
             projectDatasheet = "stsim_StateClass")

# Datasheet: Transition type
edit_project(projectInput = "Transition Type.csv",
             projectDatasheet = "stsim_TransitionType")

# Datasheet: Transition group
edit_project(projectInput = "Transition Group.csv",
             projectDatasheet = "stsim_TransitionGroup")

# Datasheet: Transition types by group
edit_project(projectInput = "Transition Types by Group.csv",
             projectDatasheet = "stsim_TransitionTypeGroup")

# Datasheet: Transition simulation groups
edit_project(projectInput = "Transition Simulation Groups.csv",
             projectDatasheet = "stsim_TransitionSimulationGroup")

# Datasheet: State attribute type
edit_project(projectInput = "State Attribute Type.csv",
             projectDatasheet = "stsim_StateAttributeType")

# Datasheet: Transition attribute type
edit_project(projectInput = "Transition Attribute Type.csv",
             projectDatasheet = "stsim_TransitionAttributeType")



# Sub-scenarios ----------------------------------------------------------------

# Create folder
subscenarioFolder <- folder(stconnectProject, folder = "Sub-Scenarios")


# 00 Pipeline ------------------------------------

# Create nested folder
pipelineFolder <- folder(stconnectProject, folder = "00 Pipeline",
                         parentFolder = subscenarioFolder)

# Landscape change & Habitat Suitability transformers
targetScenario <- edit_subscenario(
  scenarioName = "Pipeline: Landscape Change, Habitat Suitability",
  scenarioInput = "Pipeline - Landscape Change, Habitat Suitability.csv",
  scenarioDatasheet = "core_Pipeline")
# Move scenario into folder
folderId(targetScenario) <- folderId(folder(stconnectProject, "00 Pipeline"))

# Circuitscape transformer
targetScenario <- edit_subscenario(
  scenarioName = "Pipeline: Circuitscape",
  scenarioInput = "Pipeline - Circuitscape.csv",
  scenarioDatasheet = "core_Pipeline")
# Move scenario into folder
folderId(targetScenario) <- folderId(folder(stconnectProject, "00 Pipeline"))


# 01 Run control ---------------------------------

# Create nested folder
runcontrolFolder <- folder(stconnectProject, folder = "01 Run Control",
                           parentFolder = subscenarioFolder)

# Create & edit sub-scenario
# 2010-2060, 10 iterations
targetScenario <- edit_subscenario(
  scenarioName = "Run Control: Spatial; 2010 - 2060; 10 iterations",
  scenarioInput = "Run Control - 10 iterations 50 timesteps.csv",
  scenarioDatasheet = "stsim_RunControl")
# Move scenario into folder
folderId(targetScenario) <- folderId(runcontrolFolder)


# 02 Landscape change ----------------------------

# Create nested folder
landscapeChangeFolder <- folder(stconnectProject, folder = "02 Landscape Change",
                                parentFolder = subscenarioFolder)


# 01 Transition pathways ----------

# Create nested sub-folder
transitionPathwaysFolder <- folder(stconnectProject, 
                                   folder = "01 Transition Pathways",
                                   parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
# Datasheet: States
targetScenario <- edit_subscenario(
  scenarioName = "Transition Pathways: LULC; Forest Composition x Age; TST Min 20",
  scenarioInput = "Transition Pathways - States.csv",
  scenarioDatasheet = "stsim_DeterministicTransition")
# Datasheet: Transitions
targetScenario <- edit_subscenario(
  scenarioName = "Transition Pathways: LULC; Forest Composition x Age; TST Min 20",
  scenarioInput = "Transition Pathways - Transitions.csv",
  scenarioDatasheet = "stsim_Transition")
# Move scenario into folder
folderId(targetScenario) <- folderId(transitionPathwaysFolder)


# 02 Initial conditions -----------

# Create nested sub-folder
initialConditionsFolder <- folder(stconnectProject, 
                                  folder = "02 Initial Conditions",
                                  parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
targetScenario <- edit_subscenario(
  scenarioName = "Initial Conditions: Spatial 90m; Monteregie",
  scenarioInput = "Initial Conditions Spatial Files.csv",
  scenarioDatasheet = "stsim_InitialConditionsSpatial")
# Move scenario into folder
folderId(targetScenario) <- folderId(initialConditionsFolder)


# 03 Transition targets -----------

# Create nested sub-folder
transitionTargetsFolder <- folder(stconnectProject, 
                                  folder = "03 Transition Targets",
                                  parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
# Historic
targetScenario <- edit_subscenario(
  scenarioName = "Transition Targets: LULC Historic 1990 - 2010",
  scenarioInput = "Transition Targets - Historic.csv",
  scenarioDatasheet = "stsim_TransitionTarget")
# Move scenario into folder
folderId(targetScenario) <- folderId(transitionTargetsFolder)


# 04 Output options ---------------

# Create nested sub-folder
outputOptionsFolder <- folder(stconnectProject, 
                              folder = "04 Output Options",
                              parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
# Datasheet: Options
targetScenario <- edit_subscenario(
  scenarioName = "Output Options",
  scenarioInput = "Output Options.csv",
  scenarioDatasheet = "stsim_OutputOptions")
# Datasheet: Spatial
targetScenario <- edit_subscenario(
  scenarioName = "Output Options",
  scenarioInput = "Output Options Spatial.csv",
  scenarioDatasheet = "stsim_OutputOptionsSpatial")
# Datasheet: Spatial Averages
targetScenario <- edit_subscenario(
  scenarioName = "Output Options",
  scenarioInput = "Output Options Spatial Averages.csv",
  scenarioDatasheet = "stsim_OutputOptionsSpatialAverage")

# Move scenario into folder
folderId(targetScenario) <- folderId(outputOptionsFolder)


# 05 Transition Multipliers -------

# Create nested sub-folder
transitionMultiplierFolder <- folder(stconnectProject, 
                                     folder = "05 Transition Multipliers",
                                     parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
targetScenario <- edit_subscenario(
  scenarioName = "Transition Multipliers: Climate Baseline",
  scenarioInput = "Transition Multipliers.csv",
  scenarioDatasheet = "stsim_TransitionMultiplierValue")

# Move scenario into folder
folderId(targetScenario) <- folderId(transitionMultiplierFolder)


# 06 Transition Spatial Multiplier ---

# Create nested sub-folder
spatialMultiplierFolder <- folder(stconnectProject,
                                  folder = "06 Transition Spatial Multiplier",
                                  parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
# Agnostic
targetScenario <- edit_subscenario(
  scenarioName = "Transition Spatial Multiplier: Agnostic",
  scenarioInput = "Transition Spatial Multipliers - Agnostic.csv",
  scenarioDatasheet = "stsim_TransitionSpatialMultiplier")
# Move scenario into folder
folderId(targetScenario) <- folderId(spatialMultiplierFolder)
# Zonation
targetScenario <- edit_subscenario(
  scenarioName = "Transition Spatial Multiplier: Zonation",
  scenarioInput = "Transition Spatial Multipliers - Zonation.csv",
  scenarioDatasheet = "stsim_TransitionSpatialMultiplier")
# Move scenario into folder
folderId(targetScenario) <- folderId(spatialMultiplierFolder)
# Prioritizr
targetScenario <- edit_subscenario(
  scenarioName = "Transition Spatial Multiplier: Prioritizr",
  scenarioInput = "Transition Spatial Multipliers - Prioritizr.csv",
  scenarioDatasheet = "stsim_TransitionSpatialMultiplier")
# Move scenario into folder
folderId(targetScenario) <- folderId(spatialMultiplierFolder)
# Random
targetScenario <- edit_subscenario(
  scenarioName = "Transition Spatial Multiplier: Random",
  scenarioInput = "Transition Spatial Multipliers - Random.csv",
  scenarioDatasheet = "stsim_TransitionSpatialMultiplier")
# Move scenario into folder
folderId(targetScenario) <- folderId(spatialMultiplierFolder)


# 07 Transition Adjacency Multiplier ---

# Create nested sub-folder
adjacencyMultiplierFolder <- folder(stconnectProject, 
                                    folder = "07 Transition Adjacency Multiplier",
                                    parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
# Datasheet: Multipliers
targetScenario <- edit_subscenario(
  scenarioName = "Transition Adjacency Multiplier: Agriculture and Urban",
  scenarioInput = "Transition Adjacency Multipliers.csv",
  scenarioDatasheet = "stsim_TransitionAdjacencyMultiplier")
# Datasheet: Settings
targetScenario <- edit_subscenario(
  scenarioName = "Transition Adjacency Multiplier: Agriculture and Urban",
  scenarioInput = "Transition Adjacency Setting.csv",
  scenarioDatasheet = "stsim_TransitionAdjacencySetting")
# Move scenario into folder
folderId(targetScenario) <- folderId(adjacencyMultiplierFolder)


# 08 Attributes -------------------

# Create nested sub-folder
attributesFolder <- folder(stconnectProject,
                           folder = "08 Attributes",
                           parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
# State values
targetScenario <- edit_subscenario(
  scenarioName = "State Attribute Values",
  scenarioInput = "State Attribute Values.csv",
  scenarioDatasheet = "stsim_StateAttributeValue")
# Move scenario into folder
folderId(targetScenario) <- folderId(attributesFolder)

# Transition values
targetScenario <- edit_subscenario(
  scenarioName = "Transition Attribute Values",
  scenarioInput = "Transition Attribute Values.csv",
  scenarioDatasheet = "stsim_TransitionAttributeValue")
# Move scenario into folder
folderId(targetScenario) <- folderId(attributesFolder)


# 09 Transition Size --------------

# Create nested sub-folder
transitionSizeFolder <- folder(stconnectProject, 
                               folder = "09 Transition Size",
                               parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
# Datasheet: Distribution
targetScenario <- edit_subscenario(
  scenarioName = "Transition Size",
  scenarioInput = "Transition Size Distribution.csv",
  scenarioDatasheet = "stsim_TransitionSizeDistribution")
# Datasheet: Prioritization 
targetScenario <- edit_subscenario(
  scenarioName = "Transition Size",
  scenarioInput = "Transition Size Prioritization.csv",
  scenarioDatasheet = "stsim_TransitionSizePrioritization")

# Move scenario into folder
folderId(targetScenario) <- folderId(transitionSizeFolder)


# 10 Time Since Transition --------

# Create nested sub-folder
tstFolder <- folder(stconnectProject,
                    folder = "10 Time Since Transition",
                    parentFolder = landscapeChangeFolder)

# Create & edit sub-scenario
# Datasheet: Group
targetScenario <- edit_subscenario(
  scenarioName = "Time Since Transition: 20",
  scenarioInput = "Time Since Transition Group.csv",
  scenarioDatasheet = "stsim_TimeSinceTransitionGroup")
# Datasheet: Randomize
targetScenario <- edit_subscenario(
  scenarioName = "Time Since Transition: 20",
  scenarioInput = "Initial TST Randomize.csv",
  scenarioDatasheet = "stsim_TimeSinceTransitionRandomize")

# Move scenario into folder
folderId(targetScenario) <- folderId(tstFolder)



# 03 Habitat -------------------------------------

# Create nested folder
habitatFolder <- folder(stconnectProject, folder = "03 Habitat",
                        parentFolder = "Sub-Scenarios")


# 01 Habitat Suitability ----------

# Create nested sub-folder
habitatSuitabilityFolder <- folder(stconnectProject, 
                                   folder = "01 Habitat Suitability",
                                   parentFolder = "03 Habitat")

# Create & edit sub-scenario
targetScenario <- edit_subscenario(
  scenarioName = "Habitat Suitability: Basic Landcover and Forest Age - 14 species",
  scenarioInput = "Habitat Suitability.csv",
  scenarioDatasheet = "stconnect_HSHabitatSuitability")

# Move scenario into folder
folderId(targetScenario) <- folderId(habitatSuitabilityFolder)


# 02 Habitat Patch ----------------

# Create nested sub-folder
habitatPatchFolder <- folder(stconnectProject, 
                             folder = "02 Habitat Patch",
                             parentFolder = "03 Habitat")

# Create & edit sub-scenario
targetScenario <- edit_subscenario(
  scenarioName = "Habitat Patch: 14 species",
  scenarioInput = "Habitat Patch.csv",
  scenarioDatasheet = "stconnect_HSHabitatPatch")

# Move scenario into folder
folderId(targetScenario) <- folderId(habitatPatchFolder)



# 04 Circuitscape --------------------------------

# Create nested folder
circuitscapeFolder <- folder(stconnectProject, folder = "04 Circuitscape",
                             parentFolder = "Sub-Scenarios")

# 01 Resistance -------------------

# Create nested sub-folder
resistanceFolder <- folder(stconnectProject,
                           folder = "01 Resistance",
                           parentFolder = "04 Circuitscape")

# Create & edit sub-scenario
targetScenario <- edit_subscenario(
  scenarioName = "Resistance: 14 species",
  scenarioInput = "Resistance.csv",
  scenarioDatasheet = "stconnect_CCResistance")

# Move scenario into folder
folderId(targetScenario) <- folderId(resistanceFolder)



# Sensitivity scenarios --------------------------------------------------------

# Create folder
sensitivityScenarioFolder <- folder(stconnectProject, 
                                    folder = "Sensitivity Scenarios")

# Create nested sub-folders
agnosticFolder <- folder(stconnectProject,
                         folder = "Agnostic",
                         parentFolder = "Sensitivity Scenarios")
zonationFolder <- folder(stconnectProject,
                         folder = "Zonation",
                         parentFolder = "Sensitivity Scenarios")
prioritizrFolder <- folder(stconnectProject,
                       folder = "Prioritizr",
                       parentFolder = "Sensitivity Scenarios")
randomFolder <- folder(stconnectProject,
                       folder = "Random",
                       parentFolder = "Sensitivity Scenarios")

# Landscape Change & Habitat Suitability
targetScenario <- create_fullscenario(
  scenarioName = "Landscape Change & Habitat Suitability - Agnostic",
  destinationFolder = agnosticFolder,
  dependenciesCore = c(1, 3:8, 9, 13:19),
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Landscape Change & Habitat Suitability - Zonation",
  destinationFolder = zonationFolder,
  dependenciesCore = c(1, 3:8, 10, 13:19),
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Landscape Change & Habitat Suitability - Prioritizr",
  destinationFolder = prioritizrFolder,
  dependenciesCore = c(1, 3:8, 11, 13:19),
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Landscape Change & Habitat Suitability - Random",
  destinationFolder = randomFolder,
  dependenciesCore = c(1, 3:8, 12, 13:19),
  dependenciesUnique = NA)

# Run scenario
# NOTE: The scenarios were ran manually in rsyncrosim so that the Habitat 
#       Suitability transformer could be run only for 2060 and for the three key
#       species.

# Circuitscape -----------------------------------
# NOTE: Confirm the ID of the results scenario (i.e., 21 & 39 below)

# Agnostic
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 90 m resolution - Agnostic",
  destinationFolder = NA,
  dependenciesCore = c(20, 25, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 180 m resolution - Agnostic",
  destinationFolder = NA,
  dependenciesCore = c(20, 25, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution - Agnostic",
  destinationFolder = NA,
  dependenciesCore = c(20, 25, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 540 m resolution - Agnostic",
  destinationFolder = NA,
  dependenciesCore = c(20, 25, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 50 nodes - Agnostic",
  destinationFolder = NA,
  dependenciesCore = c(20, 25, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 35 nodes - Agnostic",
  destinationFolder = NA,
  dependenciesCore = c(20, 25, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 2 nodes - Agnostic",
  destinationFolder = NA,
  dependenciesCore = c(20, 25, 2), 
  dependenciesUnique = NA)

# Zonation
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 90 m resolution - Zonation",
  destinationFolder = zonationFolder,
  dependenciesCore = c(20, 26, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 180 m resolution - Zonation",
  destinationFolder = zonationFolder,
  dependenciesCore = c(20, 26, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution - Zonation",
  destinationFolder = zonationFolder,
  dependenciesCore = c(20, 26, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 540 m resolution - Zonation",
  destinationFolder = zonationFolder,
  dependenciesCore = c(20, 26, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 50 nodes - Zonation",
  destinationFolder = NA,
  dependenciesCore = c(20, 26, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 35 nodes - Zonation",
  destinationFolder = NA,
  dependenciesCore = c(20, 26, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 2 nodes - Zonation",
  destinationFolder = NA,
  dependenciesCore = c(20, 26, 2), 
  dependenciesUnique = NA)

# Random
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 90 m resolution - Random",
  destinationFolder = randomFolder,
  dependenciesCore = c(20, 28, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 180 m resolution - Random",
  destinationFolder = randomFolder,
  dependenciesCore = c(20, 28, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution - Random",
  destinationFolder = randomFolder,
  dependenciesCore = c(20, 28, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 540 m resolution - Random",
  destinationFolder = randomFolder,
  dependenciesCore = c(20, 28, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 50 nodes - Random",
  destinationFolder = NA,
  dependenciesCore = c(20, 28, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 35 nodes - Random",
  destinationFolder = NA,
  dependenciesCore = c(20, 28, 2), 
  dependenciesUnique = NA)
targetScenario <- create_fullscenario(
  scenarioName = "Circuitscape: 270 m resolution, 2 nodes - Random",
  destinationFolder = NA,
  dependenciesCore = c(20, 28, 2), 
  dependenciesUnique = NA)

# NOTE: The stconnect package does not have resolution as an input. Therefore,
#       the scenarios above were ran manually in rsyncrosim so that the 
#       resistance maps were aggregated to the target resolution and 
#       disaggregated to the original resolution.



# Backup library ---------------------------------------------------------------

# Add output folders to backup
backupDS <- datasheet(stconnectLibrary,
                      name = "core_Backup")
backupDS$IncludeOutput <- TRUE
saveDatasheet(ssimObject = stconnectLibrary, 
              data = backupDS, 
              name = "core_Backup")

# Create backup of current library version
backup(stconnectLibrary)
